feat(core): debug & silly-level logfiles
We now output all debug-level and silly-level logs into logfiles in the
project's `.garden/logs` directory.

The names of these logfiles include the command name, log level and the
time the command was executed (as an ISO timestamp).

Logfiles older than 7 days are also automatically cleaned up by this
logic.

This will be useful for gathering debug information when something
goes wrong in a Garden command that was run at a lower (less verbose)
log level.